+++ src/components/home/PartnerSection.tsx
